+++ skorch/net.py
@@ -30,6 +30,7 @@ from skorch.utils import get_map_location
 from skorch.utils import is_dataset
 from skorch.utils import noop
 from skorch.utils import params_for
+from skorch.utils import to_device
 from skorch.utils import to_numpy
 from skorch.utils import to_tensor
 from skorch.utils import train_loss_score
@@ -911,10 +912,7 @@ class NeuralNet:
         for data in iterator:
             Xi = unpack_data(data)[0]
             yp = self.evaluation_step(Xi, training=training)
-            if isinstance(yp, tuple):
-                yield tuple(n.to(device) for n in yp)
-            else:
-                yield yp.to(device)
+            yield to_device(yp, device=device)
 
     def forward(self, X, training=False, device='cpu'):
         """Gather and concatenate the output from forward call with