+++ CHANGES.md
@@ -11,8 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- When using caching in scoring callbacks, no longer uselessly iterate over the data; this can save time if iteration is slow (#552, #557)
+
 ### Fixed
 
+- Make skorch compatible with sklearn 0.22
 
 ## [0.7.0] - 2019-11-29